data слой + набросок экрана для pizza

+++ app/src/main/java/com/kicker721/pigeonpizza/pizza/domain/repository/CartRepository.kt
@@ -9,5 +9,5 @@ interface CartRepository {
 
     fun getItems(): List<CartItem>
 
-    fun getTotalPrice(): Long
+    fun getTotalPrice(): Int
 }
